x86/virt/tdx: Add skeleton for detecting and initializing TDX on demand
The TDX module is essentially a CPU-attested software module running
in the new SEAM mode to protect VMs from malicious host and certain
physical attacks.  The TDX module implements the functions to build,
tear down and start execution of protected VMs called Trusted Domains
(TD).  Before the TDX module can be used to create and run TD guests,
it must be loaded into the SEAMRR and properly initialized.  The TDX
module is expected to be loaded by BIOS before booting to the kernel,
and the kernel is expected to detect and initialize it, using the
SEAMCALLs defined by TDX architecture.

The TDX module can be initialized only once in its lifetime.  Instead
of always initializing it at boot time, this implementation chooses an
on-demand approach to initialize TDX until there is a real need (e.g
when requested by KVM).  This avoids consuming the memory that must be
allocated by kernel and given to the TDX module as metadata (~1/256th of
the TDX-usable memory), and also saves the time of initializing the TDX
module (and the metadata) when TDX is not used at all.

Introduce two placeholders tdx_detect() and tdx_init() to detect and
initialize the TDX module on demand, with a state machine introduced to
orchestrate the entire process (in case of multiple callers).

To start with, tdx_detect() checks SEAMRR and TDX private KeyIDs.  The
TDX module is reported as not-loaded if either SEAMRR is not enabled, or
there is no enough TDX private KeyIDs to create any TD guest.  The TDX
module itself requires one global TDX private KeyID to crypto protect
its metadata.

And tdx_init() is currently empty.  The TDX module will be initialized
in multi-steps defined by the TDX architecture:

  1) Global initialization;
  2) Logical-CPU scope initialization;
  3) Enumerate the TDX module capabilities and platform configuration;
  4) Configure the TDX module about usable memory ranges and global
     KeyID information;
  5) Package-scope configuration for the global KeyID;
  6) Initialize usable memory ranges based on 4).

The TDX module can also be shut down at any time during module's
lifetime.  In case of any error during the initialization process,
shut down the module.  It's pointless to leave the module in some
intermediate state during the initialization.

SEAMCALLs used in the above steps (including shutting down TDX module)
require SEAMRR being enabled and CPU being already in VMX operation
(VMXON has been done).  So far KVM is the only user of TDX, and KVM
already puts all online cpus into VMX operation.  Handling VMXON isn't
something trivial, so this implementation doesn't handle VMXON in both
tdx_detect() and tdx_init(), but requires the user of TDX to put all
cpus into VMX operation before calling them.

Signed-off-by: Kai Huang <kai.huang@intel.com>